WIP with nasty code, but kinda working

+++ engine/View.hpp
@@ -13,8 +13,8 @@ class View {
         virtual void onHideView() {}
         virtual void onShowView() {}
 
-        virtual void onUpdate(double time_since_start, float dt) = 0;
-        virtual void onDraw(double time_since_start, float dt) = 0;
+        virtual void onUpdate(double time_since_start, double dt) = 0;
+        virtual void onDraw(double time_since_start, double dt) = 0;
 
         virtual void onKeyPress(int key) {}
         virtual void onKeyRelease(int key) {}
@@ -42,12 +42,12 @@ class DefaultView: public View {
         {
         }
 
-        void onUpdate(double time_since_start, float dt)
+        void onUpdate(double time_since_start, double dt)
         {
             // physic, logic ...
         }
 
-        void onDraw(double time_since_start, float dt)
+        void onDraw(double time_since_start, double dt)
         {
             // glClearColor(0.2f, 0.2f, 0.2f, 1.0f);
             // glClear(GL_COLOR_BUFFER_BIT);
